% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-projection.R
\name{autoplot.projection}
\alias{autoplot.projection}
\alias{plot.projection}
\title{Visualize the projected PDF distribution}
\usage{
\method{autoplot}{projection}(object)

\method{plot}{projection}(object)
}
\arguments{
\item{object}{An object of the \code{projection} class.}
}
\value{
A \code{ggplot2} object.
}
\description{
Visualize the projected PDF distribution
}
\examples{
library(ggplot2)

x <- as.matrix(diff(log(EuStockMarkets[ , 1])))

proj3  <- project_t(x, .horizon = 3, .n = 2000)
proj5  <- project_t(x, .horizon = 5, .n = 2000)
proj10 <- project_t(x, .horizon = 10, .n = 2000)

autoplot(proj3)
autoplot(proj5)
autoplot(proj10)
}
