% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_projection.R
\name{new_projection}
\alias{new_projection}
\title{COnstructor for the `projection` class.}
\usage{
new_projection(.name, .projection, .horizon)
}
\description{
COnstructor for the `projection` class.
}
\keyword{internal}
