% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_model.R
\name{project_garch_model}
\alias{project_garch_model}
\alias{project_garch_model.default}
\alias{project_garch_model.tbl}
\alias{project_garch_model.xts}
\alias{project_garch_model.matrix}
\title{Projection for Conditional GARCH Models}
\usage{
project_garch_model(.invariant, .horizon, .n = 10000, .model, ...)

\method{project_garch_model}{default}(.invariant, .horizon, .n = 10000, .model, ...)

\method{project_garch_model}{tbl}(.invariant, .horizon, .n = 10000, .model, ...)

\method{project_garch_model}{xts}(.invariant, .horizon, .n = 10000, .model, ...)

\method{project_garch_model}{matrix}(.invariant, .horizon, .n = 10000, .model, ...)
}
\arguments{
\item{.invariant}{An univariate time-series.}

\item{.horizon}{A \code{double}. The horizon in which the distribution should be projected.}

\item{.n}{A \code{double}. The number of simulations that should be performed
in order to compute the pdf distribution at the horizon.}

\item{.model}{An object from the `uGARCHspec` class. See \code{\link[rugarch]{ugarchspec}}
for details.}

\item{...}{Additional parameters to be passed thought \code{\link[rugarch]{ugarchspec}}.}
}
\value{
A tidy \code{tibble} of the \code{next_step} class.
}
\description{
Fits and simulates conditional paths for GARCH models up to a given future horizon.
}
\examples{
x <- matrix(diff(log(EuStockMarkets[ , c(1, 3)])), ncol = 2)
colnames(x) <- c("DAX", "CAC")
project_garch_model(.invariant    = x,
                    .horizon      = 21,
                    .n            = 5,
                    .model        = rugarch::ugarchspec())
}
