% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_dist.R
\name{project_ghyp}
\alias{project_ghyp}
\alias{project_ghyp.default}
\alias{project_ghyp.tbl}
\alias{project_ghyp.xts}
\alias{project_ghyp.matrix}
\alias{project_ghyp.numeric}
\title{Project of the Generalized Hyperbolic Distribution}
\usage{
project_ghyp(.invariant, .horizon = 12, .n, .symmetric = FALSE)

\method{project_ghyp}{default}(.invariant, .horizon = 12, .n, .symmetric = FALSE)

\method{project_ghyp}{tbl}(.invariant, .horizon = 12, .n, .symmetric = FALSE)

\method{project_ghyp}{xts}(.invariant, .horizon = 12, .n, .symmetric = FALSE)

\method{project_ghyp}{matrix}(.invariant, .horizon = 12, .n, .symmetric = FALSE)

\method{project_ghyp}{numeric}(.invariant, .horizon = 12, .n, .symmetric = FALSE)
}
\arguments{
\item{.invariant}{A tabular (non-tidy) time-series.}

\item{.horizon}{A number. The horizon in which the distribution should be projected.}

\item{.n}{An \code{integer} with the number of interpolations (Monte-Carlo simulations)
to be conducted.}

\item{.symmetric}{A flag. Should the estimated distribution be symmetric?
Defaults to \code{FALSE}.}
}
\value{
A \code{tibble} with \code{.n} rows.
}
\description{
This function uses the fourier transform operation to project the GHD into the
future.
}
\examples{
\donttest{
x <- matrix(diff(log(EuStockMarkets[ , 1])))
colnames(x) <- "DAX"

project_ghyp(x, .horizon = 12, .n = 10)
}
}
\references{
Meucci, A., 2005. Risk and Asset Allocation. Springer.
}
