% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_dist.R
\name{project_hyp}
\alias{project_hyp}
\alias{project_hyp.default}
\alias{project_hyp.tbl}
\alias{project_hyp.xts}
\alias{project_hyp.matrix}
\alias{project_hyp.numeric}
\title{Project the Hyperbolic Distribution}
\usage{
project_hyp(.invariant, .horizon = 12, .n, .symmetric = FALSE)

\method{project_hyp}{default}(.invariant, .horizon = 12, .n, .symmetric = FALSE)

\method{project_hyp}{tbl}(.invariant, .horizon = 12, .n, .symmetric = FALSE)

\method{project_hyp}{xts}(.invariant, .horizon = 12, .n, .symmetric = FALSE)

\method{project_hyp}{matrix}(.invariant, .horizon = 12, .n, .symmetric = FALSE)

\method{project_hyp}{numeric}(.invariant, .horizon = 12, .n, .symmetric = FALSE)
}
\arguments{
\item{.invariant}{A tabular (non-tidy) time-series.}

\item{.horizon}{A number. The horizon in which the distribution should be projected.}

\item{.n}{An \code{integer} with the number of interpolations (Monte-Carlo simulations)
to be conducted.}

\item{.symmetric}{Should fitted distribution be symmetric? The default is FALSE.}
}
\value{
A \code{tibble} with \code{.n} rows.
}
\description{
This function uses the fourier transform operation to project the HD into the
future.
}
\examples{
\donttest{
x <- matrix(diff(log(EuStockMarkets[ , 1])))
colnames(x) <- "DAX"

project_hyp(x, .horizon = 12, .n = 10)
}
}
\references{
Meucci, A., 2005. Risk and Asset Allocation. Springer.
}
