% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_model.R
\name{project_mean_model}
\alias{project_mean_model}
\alias{project_mean_model.default}
\alias{project_mean_model.tbl}
\alias{project_mean_model.xts}
\alias{project_mean_model.matrix}
\alias{project_mean_model.numeric}
\title{Projection for Conditional Mean Models}
\usage{
project_mean_model(.invariant, .horizon, .n = 10000, .model, ...)

\method{project_mean_model}{default}(.invariant, .horizon, .n = 10000, .model, ...)

\method{project_mean_model}{tbl}(.invariant, .horizon, .n = 10000, .model, ...)

\method{project_mean_model}{xts}(.invariant, .horizon, .n = 10000, .model, ...)

\method{project_mean_model}{matrix}(.invariant, .horizon, .n = 10000, .model, ...)

\method{project_mean_model}{numeric}(.invariant, .horizon, .n = 10000, .model, ...)
}
\arguments{
\item{.invariant}{An univariate time-series.}

\item{.horizon}{A \code{double}. The horizon in which the distribution should be projected.}

\item{.n}{A \code{double}. The number of simulations that should be performed
in order to compute the pdf distribution at the horizon.}

\item{.model}{A \code{character}. One of: \code{\link[forecast]{ets}},
\code{\link[stats]{ar}}, \code{\link[stats]{arima}}, \code{\link[forecast]{auto.arima}} or
\code{\link[forecast]{nnetar}}.}

\item{...}{Additional parameters to be passed thought `.model`.}
}
\value{
A tidy \code{tibble} of the \code{next_step} class.
}
\description{
Fits and simulates conditional paths for mean models up to a given future horizon.
}
\examples{
x <- matrix(diff(log(EuStockMarkets[ , c(1, 2)])), ncol = 2)
colnames(x) <- c("DAX", "SMI")
project_mean_model(.invariant    = x,
                   .horizon      = 21,
                   .n            = 10,
                   .model        = "arima") # random-walk

project_mean_model(.invariant    = x,
                   .horizon      = 21,
                   .n            = 10,
                   .model        = "nnetar") # neural-networks
}
