% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_dist.R
\name{project_nig}
\alias{project_nig}
\alias{project_nig.default}
\alias{project_nig.tbl}
\alias{project_nig.xts}
\alias{project_nig.matrix}
\alias{project_nig.numeric}
\title{Project the Normal Inverse Gaussian Distribution}
\usage{
project_nig(.invariant, .horizon = 12, .n, .symmetric = FALSE)

\method{project_nig}{default}(.invariant, .horizon = 12, .n, .symmetric = FALSE)

\method{project_nig}{tbl}(.invariant, .horizon = 12, .n, .symmetric = FALSE)

\method{project_nig}{xts}(.invariant, .horizon = 12, .n, .symmetric = FALSE)

\method{project_nig}{matrix}(.invariant, .horizon = 12, .n, .symmetric = FALSE)

\method{project_nig}{numeric}(.invariant, .horizon = 12, .n, .symmetric = FALSE)
}
\arguments{
\item{.invariant}{A tabular (non-tidy) time-series.}

\item{.horizon}{A number. The horizon in which the distribution should be projected.}

\item{.n}{An \code{integer} with the number of interpolations (Monte-Carlo simulations)
to be conducted.}

\item{.symmetric}{Should fitted distribution be symmetric? The default is FALSE.}
}
\value{
A \code{tibble} with \code{.n} rows.
}
\description{
This function uses the fourier transform operation to project the NIG into the
future.
}
\examples{
\donttest{
x <- matrix(diff(log(EuStockMarkets[ , 1])))
colnames(x) <- "DAX"

project_nig(x, .horizon = 12, .n = 10)
}
}
\references{
Meucci, A., 2005. Risk and Asset Allocation. Springer.
}
