% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_dist.R
\name{project_np}
\alias{project_np}
\alias{project_np.default}
\alias{project_np.tbl}
\alias{project_np.xts}
\alias{project_np.matrix}
\alias{project_np.numeric}
\title{Non-Parametric Projection}
\usage{
project_np(.invariant, .horizon = 5, .n)

\method{project_np}{default}(.invariant, .horizon = 5, .n)

\method{project_np}{tbl}(.invariant, .horizon = 5, .n)

\method{project_np}{xts}(.invariant, .horizon = 5, .n)

\method{project_np}{matrix}(.invariant, .horizon = 5, .n)

\method{project_np}{numeric}(.invariant, .horizon = 5, .n)
}
\arguments{
\item{.invariant}{A tabular (non-tidy) time-series.}

\item{.horizon}{A number. The horizon in which the distribution should be projected.}

\item{.n}{An \code{integer} with the number of interpolations (Monte-Carlo simulations)
to be conducted.}
}
\value{
A \code{tibble} with \code{.n} rows.
}
\description{
Applies the Discrete Fourier Transform to project distributions non-parametrically
into the future.
}
\examples{
\donttest{
# log returns
x <- matrix(diff(log(EuStockMarkets[ , "FTSE", drop = FALSE])))
colnames(x) <- "FTSE"

project_np(x, .horizon = 12, .n = 10)
}
}
\references{
- Meucci, A., 2005. Risk and Asset Allocation. Springer.

- Attilio Meucci (2021). Exercises in Advanced Risk and Portfolio Management
(https://www.mathworks.com/matlabcentral/fileexchange/25010-exercises-in-advanced-risk-and-portfolio-management),
MATLAB Central File Exchange. Retrieved August 4, 2021.
}
\seealso{
\code{\link{project_t_uv}}
}
