% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_dist.R
\name{project_t}
\alias{project_t}
\alias{project_t.default}
\alias{project_t.tbl}
\alias{project_t.xts}
\alias{project_t.matrix}
\alias{project_t.numeric}
\title{Student-t Projection}
\usage{
project_t(.invariant, .horizon = 12, .n)

\method{project_t}{default}(.invariant, .horizon = 12, .n)

\method{project_t}{tbl}(.invariant, .horizon = 12, .n)

\method{project_t}{xts}(.invariant, .horizon = 12, .n)

\method{project_t}{matrix}(.invariant, .horizon = 12, .n)

\method{project_t}{numeric}(.invariant, .horizon = 12, .n)
}
\arguments{
\item{.invariant}{A tabular (non-tidy) time-series.}

\item{.horizon}{A number. The horizon in which the distribution should be projected.}

\item{.n}{An \code{integer} with the number of interpolations (Monte-Carlo simulations)
to be conducted.}
}
\value{
A \code{tibble} with \code{.n} rows.
}
\description{
Applies the Discrete Fourier Transform to project the t-distribution into the future.
}
\examples{
\donttest{
# log returns
x <- matrix(diff(log(EuStockMarkets[ , "FTSE", drop = FALSE])))
colnames(x) <- "FTSE"

project_t(x, .horizon = 12, .n = 10)
}
}
\references{
- Meucci, A., 2005. Risk and Asset Allocation. Springer.

- Attilio Meucci (2021). Exercises in Advanced Risk and Portfolio Management
(https://www.mathworks.com/matlabcentral/fileexchange/25010-exercises-in-advanced-risk-and-portfolio-management),
MATLAB Central File Exchange. Retrieved August 4, 2021.
}
\seealso{
\code{\link{project_np}}
}
