% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_dist_uv.R
\name{project_t_uv}
\alias{project_t_uv}
\title{Projection of univariate t distributions}
\usage{
project_t_uv(.invariant, .horizon = 12, .n)
}
\description{
Projection of univariate t distributions
}
\keyword{internal}
