% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_dist.R
\name{project_vg}
\alias{project_vg}
\alias{project_vg.default}
\alias{project_vg.tbl}
\alias{project_vg.xts}
\alias{project_vg.matrix}
\alias{project_vg.numeric}
\title{Project the Variance Gamma Distribution}
\usage{
project_vg(.invariant, .horizon = 12, .n, .symmetric = FALSE)

\method{project_vg}{default}(.invariant, .horizon = 12, .n, .symmetric = FALSE)

\method{project_vg}{tbl}(.invariant, .horizon = 12, .n, .symmetric = FALSE)

\method{project_vg}{xts}(.invariant, .horizon = 12, .n, .symmetric = FALSE)

\method{project_vg}{matrix}(.invariant, .horizon = 12, .n, .symmetric = FALSE)

\method{project_vg}{numeric}(.invariant, .horizon = 12, .n, .symmetric = FALSE)
}
\arguments{
\item{.invariant}{A tabular (non-tidy) time-series.}

\item{.horizon}{A number. The horizon in which the distribution should be projected.}

\item{.n}{An \code{integer} with the number of interpolations (Monte-Carlo simulations)
to be conducted.}

\item{.symmetric}{Should fitted distribution be symmetric? The default is FALSE.}
}
\value{
A \code{tibble} with \code{.n} rows.
}
\description{
This function uses the fourier transform operation to project the VGD into the
future.
}
\examples{
\donttest{
x <- matrix(diff(log(EuStockMarkets[ , 1])))
colnames(x) <- "DAX"

project_vg(x, .horizon = 12, .n = 10)
}
}
\references{
Meucci, A., 2005. Risk and Asset Allocation. Springer.
}
